feat(mcp): move tools into own pkg
