+++ pkg/mcp/gen/expert.go
@@ -4,49 +4,15 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/apigear-io/cli/pkg/cmd/gen"
 	"github.com/apigear-io/cli/pkg/helper"
 	"github.com/apigear-io/cli/pkg/sol"
 	"github.com/mark3labs/mcp-go/mcp"
 	"github.com/mark3labs/mcp-go/server"
 )
 
-func RegisterMCPTools(s *server.MCPServer) {
-	registerGenerateSolutionTool(s)
-	registerGenerateExpertTool(s)
-}
-
-func registerGenerateSolutionTool(s *server.MCPServer) {
-	genSolutionTool := mcp.NewTool("generateSolution",
-		mcp.WithDescription("Generate SDK based on a solution document"),
-		mcp.WithString("solution", mcp.Required(), mcp.Description("Path to solution file")),
-		mcp.WithString("force", mcp.Description("Force overwrite (true/false)")),
-		mcp.WithString("watch", mcp.Description("Watch for changes (true/false). This keeps the process running.")),
-	)
-	s.AddTool(genSolutionTool, func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		solutionPath, err := request.RequireString("solution")
-		if err != nil {
-			return mcp.NewToolResultError(err.Error()), nil
-		}
-
-		forceEnabled := false
-		if force, err := request.RequireString("force"); err == nil && force == "true" {
-			forceEnabled = true
-		}
-		watchEnabled := false
-		if watch, err := request.RequireString("watch"); err == nil && watch == "true" {
-			watchEnabled = true
-		}
-
-		err = runGenerateSolution(solutionPath, watchEnabled, forceEnabled)
-		if err != nil {
-			return mcp.NewToolResultError(err.Error()), nil
-		}
-		return mcp.NewToolResultText(fmt.Sprintf("Successfully ran code generation for solution file: %s", solutionPath)), nil
-	})
-}
-
 func registerGenerateExpertTool(s *server.MCPServer) {
-	options := &ExpertOptions{}
+	options := &gen.ExpertOptions{}
 
 	genExpertTool := mcp.NewTool("generateExpert",
 		mcp.WithDescription("Generate code using expert mode with individual generator options"),
@@ -62,33 +28,33 @@ func registerGenerateExpertTool(s *server.MCPServer) {
 		if err != nil {
 			return mcp.NewToolResultError(err.Error()), nil
 		}
-		options.inputs = append(options.inputs, input)
+		options.Inputs = append(options.Inputs, input)
 
 		output, err := request.RequireString("output")
 		if err != nil {
 			return mcp.NewToolResultError(err.Error()), nil
 		}
-		options.outputDir = output
+		options.OutputDir = output
 
 		template, err := request.RequireString("template")
 		if err != nil {
 			return mcp.NewToolResultError(err.Error()), nil
 		}
-		options.templateDir = template
+		options.TemplateDir = template
 
 		if features, err := request.RequireString("features"); err == nil && features != "" {
-			options.features = append(options.features, features)
+			options.Features = append(options.Features, features)
 		}
-		options.force = false
+		options.Force = false
 		if force, err := request.RequireString("force"); err == nil && force == "true" {
-			options.force = true
+			options.Force = true
 		}
-		options.watch = false
+		options.Watch = false
 		if watch, err := request.RequireString("watch"); err == nil && watch == "true" {
-			options.watch = true
+			options.Watch = true
 		}
 
-		doc := makeSolution(options)
+		doc := gen.MakeSolution(options)
 		if err := doc.Validate(); err != nil {
 			return mcp.NewToolResultError(fmt.Sprintf("invalid solution document: %s", err.Error())), nil
 		}
@@ -100,7 +66,7 @@ func registerGenerateExpertTool(s *server.MCPServer) {
 			return mcp.NewToolResultError(err.Error()), nil
 		}
 
-		if options.watch {
+		if options.Watch {
 			err := runner.WatchDoc(ctx, doc.RootDir, doc)
 			if err != nil {
 				cancel()
