+++ pkg/mcp/gen/root.go
@@ -1,0 +1,10 @@
+package gen
+
+import (
+	"github.com/mark3labs/mcp-go/server"
+)
+
+func RegisterMCPTools(s *server.MCPServer) {
+	registerGenerateSolutionTool(s)
+	registerGenerateExpertTool(s)
+}
