+++ pkg/mcp/spec/root.go
@@ -1,0 +1,10 @@
+package spec
+
+import (
+	"github.com/mark3labs/mcp-go/server"
+)
+
+func RegisterMCPTools(s *server.MCPServer) {
+	registerSpecCheckTool(s)
+	registerSpecShemaTool(s)
+}
