+++ pkg/mcp/tpl/root.go
@@ -1,0 +1,9 @@
+package tpl
+
+import (
+	"github.com/mark3labs/mcp-go/server"
+)
+
+func RegisterMCPTools(s *server.MCPServer) {
+	registerTemplateListTool(s)
+}
